Begun adding in code for creating the line
